geotiff tests: consolidate nodata read-side cluster (Sub-PR A)
Cluster 10 of long-tail epic #2424. Fold 8 read-side nodata files
into the existing ``xrspatial/geotiff/tests/read/test_nodata.py``
alongside the GPU helper coverage:

- ``test_masked_nodata_attr_2092.py``
- ``test_nodata_attr_aliases_1582.py``
- ``test_nodata_lifecycle_attrs_2135.py``
- ``test_nodata_lifecycle_parity_2211.py``
- ``test_nodata_nan_int_1774.py``
- ``test_nodata_no_extra_copy_1553.py``
- ``test_nodata_semantics_split_1988.py``
- ``test_helper_band_nodata_2210.py``

Tests-only. The shared ``requires_gpu`` marker from
``_helpers/markers.py`` (aliased ``_gpu_only``) replaces the
per-file ``_gpu_available`` / ``_gpu_only`` helpers. Module-level
helper functions are suffixed with the source issue number to
keep sibling sections collision-free, and the rasterio
``importorskip`` from ``test_nodata_semantics_split_1988.py`` is
localised to per-helper / per-test scope so the rest of the
module still collects when rasterio is absent.

185 tests collected (9 pre-existing GPU + 176 newly folded),
matching the baseline sum.

Cross-cutting follow-ups:

- ``test_invalid_int_nodata_rejection_2441.py`` (sibling cluster,
  not in this PR) imported ``_build_uint16_tiff`` from the
  now-deleted #1774 file; the import points at the renamed helper
  ``_build_uint16_tiff_1774`` in ``read/test_nodata.py``.
- ``docs/source/reference/release_gate_geotiff.rst``: the nodata
  round-trip row cites the consolidated files so the
  release-gate checklist-parity test still passes.
- ``docs/source/reference/geotiff.rst``: prose pointers updated.
- ``test_roundtrip_properties.py``: stale docstring pointer updated.

``CLUSTER_AUDIT_NODATA.md`` maps every ``old_file::test_name`` to
its ``new_file::test_name``. The audit file is deleted on a final
commit on this branch before merge.

Refs #2434.
